+++ lib/fishstick_dart.dart
@@ -5,6 +5,7 @@ export "commands/info/invite.dart";
 export "commands/info/info.dart";
 
 export "commands/general/autopost.dart";
+export "commands/general/color.dart";
 
 export "commands/partner/premium.dart";